[`Backend`]: refactor docs route (#148)
