+++ server/docs/__init__.py
@@ -1,0 +1,1 @@
+from .routes import router as docs_router
